abstract out identifyDeclarations
